agent_service/search/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=2.8.0
tavily-python>=0.5.0
pydantic>=2.0.0

